requirements.txt
fastapi
uvicorn[standard]
pyschlage==2024.2.0
python-dotenv
